RTPS: don't use incompletely filled in GUIDs.
Add to the GUID type a bitmask that indicates which fields have been
filled in; start it out as 0, and then set bits in it as fields get
filled in.

Do not add a type mapping object to the hash table unless the GUID is
completely filled in, and don't look for a type mapping object in the
hash table with a GUID that hasn't been completely filled in as a key.

Bug: 16642
Change-Id: I31db92238adcb2ec2d70b2650e41b14d99001908
Reviewed-on: https://code.wireshark.org/review/37537
Petri-Dish: Guy Harris <gharris@sonic.net>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <gharris@sonic.net>
(backported from commit f97e20a011144a56bf72dd92c49167d17713a909)
Reviewed-on: https://code.wireshark.org/review/37541