tn3270: improve on some of IBM's painful prose.
The "true" string in a struct true_false_string is used only if the bit
is set to 1, so "if set to logic 1" is redundant.  Just have the strings
say, respectively, what significance the bit being on ("set to logic 1")
has and what signficance the bit being off ("set to logic 0") has.

Use active voice.

Change-Id: If1fe7b35a0c85dcdb40ac348ca23a0ac5e3dc1da
Reviewed-on: https://code.wireshark.org/review/37753
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Reviewed-by: Guy Harris <gharris@sonic.net>