NIS+: give the meanings of the NIS_CALLBACK return value.
From reading GNU libc's .x file for NIS+ (which has an Oracle America
copyright, suggesting that it was originally a file from Sun, and thus
likely to be the official NIS+ rpcgen file), and from reading the way
the GNU libc code treats the return value of the return value of an
NIS_CALLBACK RPC, it appears to be a Boolean where "true" means that the
callback in question is still running and "false" means it's not
running.

Label the results as such.

(I should probably really check the Illumos source to see if it includes
the NIS+ client and server and, if so, read that, so see what Sun's code
did.)

Change-Id: I4eb430dfca3d1162972a2a750effc31f626f20bf
Reviewed-on: https://code.wireshark.org/review/37791
Reviewed-by: Guy Harris <gharris@sonic.net>