Qt: Always pop empty filter syntax
When checking a filter the status message should always be pop'ed when
having an empty filter, regardless of having the clear button or not.

This will ensure the status message is removed when removing a display
filter using DisplayFilterEdit in other dialogs than Main.

Change-Id: I3c9a4933cd0c60ab624ea1939ffafecb58b3ffd5
Reviewed-on: https://code.wireshark.org/review/38054
Petri-Dish: Stig Bjørlykke <stig@bjorlykke.org>
Tested-by: Petri Dish Buildbot
Reviewed-by: Stig Bjørlykke <stig@bjorlykke.org>