bluetooth.gaen: Adding Google/Apple Exposure Notification to bluetooth.
The Google/Apple Exposure Notification protocol is designed to aid
contact tracing efforts by allowing users to broadcast changing
identifiers, derived from longer-term (24 hour) keys; in the event that
a user receives a positive diagnosis, they upload their longer-term keys
to a key server, and all other users can use those long-term keys to
generate all the potential changing identifiers, and compare those to
their logs to determine if they were in contact with the infected user.
This protocol was developed in response to SARS-CoV-2, but is not
inherently limited to it.

This patch adds a "bluetooth.gaen" filter, with two data fields in the
periodic (changing identifier) broadcast:
 - bluetooth.gaen.rpi: The Rolling Proximity Identifier
 - bluetooth.gaen.aemd: The Associated Encrypted Metadata

Links to Protocol Documents:
 - Google: https://www.google.com/covid19/exposurenotifications/
 - Apple: https://www.apple.com/covid19/contacttracing

This change also adds the Bluetooth SIG-assigned 16-bit UUID for GAEN,
0xFD6F, to the list of Wireshark-recognizable 16-bit UUIDs.

These changes are licensed under the same license as Wireshark itself.

Change-Id: I3af14b225a35d0670433a9a89901d4d37895b3bd
Reviewed-on: https://code.wireshark.org/review/38064
Reviewed-by: Anders Broman <a.broman58@gmail.com>