Only do trailing stray character checks on FT_STRING fields.
FT_STRINGZ means "terminated by a null character", so there can't be
non-null characters following the terminating null.

FT_STRINGZPAD doesn't only mean "padded with nulls"; there are protocols
where a string that's not the full length of the part of the packet for
the string has a null terminator but isn't guaranteed to be fully padded
with nulls.  We can later add a separate type for fields where we really
*should* check that the padding is all nulls.

Change-Id: I5964817b4b847cb4db73f8ac673141052e8ef92c
Reviewed-on: https://code.wireshark.org/review/38230
Petri-Dish: Guy Harris <gharris@sonic.net>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <gharris@sonic.net>