ncp: fix a copy-and-pasteo.
In process_multivalues(), we create a protocol item for the attribute
syntax, but we don't fetch its value, and don't pass it to
print_nds_values() as the syntax argument; instead, we pass a variable
that wee initialize to 0, but never set.  (One of the disadvantages of
preemptively initializing local variables is that data flow analyzers in
compilers and static analyzers can't point out that you didn't set the
variables in question to *useful* values.)

This fixes the dissection of NDS Read replies.


(cherry picked from commit 1a410ef0b0d841c4f60936418f18ef6a5544dbd9)