packet_diagram: use a better label for FT_NONE items
FT_NONE items all have the name "Text Item" which makes it look
poor in the packet diagram.
For these fields, switch to use ->representation insead of ->name
to make these fields look more meaningful.

SMB2 is a protocol that consists of two FT_NONE expansions,
one for the Header and a second for the actual Command.
This makes packet diagram show this much nicer.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>