afs: use a simpler way to read a time field
If a time field uses a standard enconding, we can call proto_tree_add_item()
to add it to the tree. There's no need to parse the time field ourselves.

Update two places in the afs dissector where the manual parsing can
easily be replaced with a proto_tree_add_item() call.