ecmp: fix dissection of file_info request
Playing with the sample capture from bugzilla bug 10562, dissection of
packet 491 (ecmp file_info request) brought up an expert info about a
malformed packet.

The request contains a list of requested attributes. For each attribute,
only the attribute ID is part of the request. The current code tries to
dissect each attribute, this fails when we only have a list of
attribute IDs...

Add a subtree for the list of IDs (and the length of that list).

While at it, remove some unnecessary variable initializers.