Add /*build*/ to .gitignore
A common pattern is to build the source artifacts within a *build* dir.

For example, the wiki instructions for [building `wireshark` on `macOS` specify to
create a `./build`
directory](https://gitlab.com/wireshark/wireshark/-/wikis/BuildingAndInstalling#macos):
```
mkdir build && cd build && cmake ../ && make
```

So this commit adds this directory to `.gitignore` to prevent
accidentally committing build artifacts.

The specific format `/*build*/` protects against the following:
1. The leading slash makes sure only top-level directories that match
this pattern are ignored.
2. The trailing slash makes sure that only directories (and not files)
are ignored.
3. The wildcards catch all the various variations on `build` folder name
that are used...